feat: add JWT authentication and protect job routes

+++ server/src/controllers/jobs.controller.ts
@@ -1,30 +1,38 @@
 import { Request, Response } from 'express'
-import { Job, CreateJobInput } from '../types/index'
+import { CreateJobInput } from '../types/index'
 import { prisma } from '../lib/prisma'
 
 export const getAllJobs = async (
   req: Request,
   res: Response
 ): Promise<void> => {
-  const jobs = await prisma.job.findMany()
+  const user = req.userId
+
+  if (!user) {
+    res.status(401).json({ success: false, message: 'Unauthorized' })
+    return
+  }
+
+  const jobs = await prisma.job.findMany({ where: { userId: user } })
 
   res.json({
     success: true,
     data: jobs,
   })
 }
-// export const getAllJobs = (req: Request, res: Response): void => {
-//   res.json({
-//     success: true,
-//     data: jobs,
-//   })
-// }
 
 export const createJob = async (req: Request, res: Response): Promise<void> => {
   const body = req.body as CreateJobInput
+  const user = req.userId
+
+  if (!user) {
+    res.status(401).json({ success: false, message: 'Unauthorized' })
+    return
+  }
 
   const newJob = await prisma.job.create({
     data: {
+      userId: user,
       company: body.company,
       role: body.role,
       jobUrl: body.jobUrl,
@@ -41,8 +49,15 @@ export const createJob = async (req: Request, res: Response): Promise<void> => {
 
 export const deleteJob = async (req: Request, res: Response): Promise<void> => {
   const id = req.params.id as string
+  const user = req.userId
+
+  if (!user) {
+    res.status(401).json({ success: false, message: 'Unauthorized' })
+    return
+  }
+
   try {
-    await prisma.job.delete({ where: { id: id } })
+    await prisma.job.delete({ where: { id: id, userId: user } })
 
     res.json({
       success: true,
@@ -59,10 +74,16 @@ export const deleteJob = async (req: Request, res: Response): Promise<void> => {
 export const updateJob = async (req: Request, res: Response): Promise<void> => {
   const id = req.params.id as string
   const body = req.body as CreateJobInput
+  const user = req.userId
+
+  if (!user) {
+    res.status(401).json({ success: false, message: 'Unauthorized' })
+    return
+  }
 
   try {
     const updatedJob = await prisma.job.update({
-      where: { id: id },
+      where: { id: id, userId: user },
       data: { ...body },
     })
     res.json({
